% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{bamtobed}
\alias{bamtobed}
\title{Convert Bam to Bed}
\usage{
bamtobed(file, outFile = gsub("\\\\.bam", ".bed", file), filtDup = FALSE)
}
\arguments{
\item{file}{File to process.}

\item{outFile}{Name of output file.}

\item{filtDup}{Output index name}
}
\value{
Path
}
\description{
Convert Bam to Bed
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_QFCollapsed)
FqFile_QFColStpClipped <- fastx_clipper(FqFile_QFColStripped)
bam <- bowtie_align(FqFile_QFColStpClipped,myIndex)
bamtobed(bam)
}
\author{
Kathryn Rozen-Gagnon
}
