% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{bzip2}
\alias{bzip2}
\title{Wrapper function for bzip2}
\usage{
bzip2(
  fileToBzip2,
  bzip2 = "Bzip2",
  keep = TRUE,
  force = FALSE,
  small = FALSE,
  blockSize = 1,
  stderr = file.path(dirname(fileToBzip2), paste0(basename(fileToBzip2),
    "_Bzip2_stderr.txt")),
  stdout = file.path(dirname(fileToBzip2), paste0(basename(fileToBzip2),
    "_Bzip2_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileToBzip2}{path to file to bzip.}

\item{bzip2}{path to bzip2.}

\item{keep}{keep (don't delete) input files, TRUE (default) or FALSE.}

\item{force}{overwrite existing output files.}

\item{small}{use less memory (at most 2500k).}

\item{blockSize}{set block size to 100k .. 900k.}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to bzipped file.
}
\description{
Wrapper function for bzip2
}
\author{
Kathryn Rozen-Gagnon
}
