% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chimera_Functions.R
\docType{methods}
\name{chimera_joinread}
\alias{chimera_joinread}
\alias{chimera_Process}
\title{Merge read sequence with bedfile}
\usage{
chimera_Process(
  bams,
  knownMiRNAs,
  genomeIndex,
  exclude,
  bpparam = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bams}{BAM files to process.}

\item{knownMiRNAs}{knownMiRNAs FASTA file.}

\item{genomeIndex}{Full genome index.}

\item{exclude}{Names of miRNAs or sequences to remove, character vector.}

\item{file}{File to process.}

\item{outFile}{Name of output file.}

\item{filtDup}{Output index name}

\item{length_min}{minimum length required during fasta processing, integer.}

\item{length_max}{maximum length required during fasta processing, integer.}
}
\value{
Path

Path
}
\description{
Merge read sequence with bedfile
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_QFCollapsed)
FqFile_QFColStpClipped <- fastx_clipper(FqFile_QFColStripped)
bam <- bowtie_align(FqFile_QFColStpClipped,myIndex)
bamtobed(bam)
}
\author{
Kathryn Rozen-Gagnon

Kathryn Rozen-Gagnon
}
