% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{CLIP_align}
\alias{CLIP_align}
\title{Align to Genome}
\usage{
CLIP_align(
  samID,
  res_dir = NULL,
  genome_idx = NULL,
  aligner = NULL,
  samSheet = NULL,
  thread = 4,
  bwa = NULL
)
}
\arguments{
\item{samID}{sample id in the sample sheet.}

\item{res_dir}{result directory.}

\item{genome_idx}{path to indexed genome.}

\item{aligner}{Assign aligner bowtie2/bwa_mem/bwa_alb/subread/subjunc/hisat2}

\item{samSheet}{sample sheet.}

\item{thread}{thread used in the processing, default is 4}

\item{bwa}{executable bwa file path.}
}
\value{
path to sorted BAM file.
}
\description{
Align to Genome
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
}
