% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_buildIDX.R
\docType{methods}
\name{ClIP_buildIDX}
\alias{ClIP_buildIDX}
\title{Build index for referene genome}
\usage{
ClIP_buildIDX(
  ref_seq = NULL,
  res_dir = NULL,
  preFIX = NULL,
  aligner = NULL,
  thread = 4,
  bwa = NULL
)
}
\arguments{
\item{ref_seq}{reference genome}

\item{res_dir}{result direcotry}

\item{preFIX}{file name prefix for indexed genome}

\item{aligner}{Assign aligner bowtie2/subread/bwa/hisat2}

\item{thread}{threads used in processing, default is 4}

\item{bwa}{exercutable bwa file path}
}
\value{
indexed reference genome
}
\description{
Build index for referene genome
}
