% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{CLIP_buildIDX}
\alias{CLIP_buildIDX}
\title{Build index for reference genome}
\usage{
CLIP_buildIDX(
  ref_seq = NULL,
  res_dir = NULL,
  preFIX = NULL,
  aligner = NULL,
  thread = 4,
  bwa = NULL
)
}
\arguments{
\item{ref_seq}{path to reference genome.}

\item{res_dir}{result directory.}

\item{preFIX}{file name prefix for indexed genome.}

\item{aligner}{Assign aligner bowtie2/subread/bwa/hisat2.}

\item{thread}{threads used in processing, default is 4.}

\item{bwa}{executable bwa file path.}
}
\value{
indexed reference genome.
}
\description{
Build index for reference genome
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
}
