% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{compress}
\alias{compress}
\title{Wrapper function for gzip, bzip}
\usage{
compress(
  fileToCompress,
  keep = TRUE,
  methodCompress = "gzip",
  overwrite = FALSE
)
}
\arguments{
\item{fileToCompress}{path to file to compress.}

\item{keep}{keep (don't delete) input files, TRUE (default) or FALSE.}

\item{methodCompress}{method to use for compression.}

\item{overwrite}{overwrite existing output files, TRUE or FALSE (default).}
}
\value{
path to zipped file.
}
\description{
Wrapper function for gzip, bzip
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
decom <- decompress(testFQ,overwrite=TRUE)
com <- compress(decom,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
