% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_cims}
\alias{ctk_cims}
\title{Wrapper function for ctk's CIMS}
\usage{
ctk_cims(
  filesToRun,
  mutationBedFile,
  outFile = paste(file_path_sans_ext(fileToRun), "CIMS", "txt", sep = "."),
  sb = "CIMS.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  mutationSize = 1,
  permutations = 5,
  trackMutationPos = FALSE,
  noSparseCorrect = FALSE,
  FDR = 1,
  mfr = 0,
  cacheDir = paste0(filesToRun, "_cache", paste(sample(letters, 10), collapse = "")),
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{mutationBedFile}{Mutation BED file.}

\item{outFile}{Output file}

\item{sb}{Path to fastx_barcode_splitter.pl from FastX toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{bigFile}{big file}

\item{mutationSize}{mutation size}

\item{permutations}{number of iterations for permutation}

\item{trackMutationPos}{track mutation position relative to read start}

\item{noSparseCorrect}{no sparcity correction}

\item{FDR}{threshold of FDR}

\item{mfr}{threshold of m-over-k-ratio}

\item{cacheDir}{Name for cache directory.}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's CIMS
}
\examples{
\dontrun{
mutations <- system.file("extdata/BrdU.Fox.pool.tag.uniq.mutation.small.txt",package="CLIPflexR")
delBed <- ctk_getMutationType(mutations)
ctk_cims("~/Downloads/uniq_tags_mutations/Fox.pool.tag.uniq.rgb.bed",delBed,verbose=TRUE)
}
}
\author{
Kathryn Rozen-Gagnon
}
