% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_cits}
\alias{ctk_cits}
\title{Wrapper function for ctk's CITS}
\usage{
ctk_cits(
  filesToRun,
  outFile = paste(file_path_sans_ext(fileToRun), "CITS", "bed", sep = "."),
  sb = "CITS.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  pCutOff = 0.01,
  multiTest = TRUE,
  gap = 25,
  cacheDir = paste0(filesToRun, "_cache", paste(sample(letters, 10), collapse = "")),
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun), "_cits_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun), "_cits_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{outFile}{output file (BED).}

\item{sb}{path to CITS.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{bigFile}{big file, TRUE or FALSE (default).}

\item{pCutOff}{p-value cut off (default is 0.01).}

\item{multiTest}{perform multiple testing correction, TRUE (default) or FALSE.}

\item{gap}{gap size (default is 25).}

\item{cacheDir}{name for cache directory.}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
path to CITS bed file.
}
\description{
Wrapper function for ctk's CITS
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_Col<- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
}
\author{
Kathryn Rozen-Gagnon
}
