% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_fastqFilter}
\alias{ctk_fastqFilter}
\title{Wrapper function for ctk's fastq_filter}
\usage{
ctk_fastqFilter(
  filesToRun,
  outFile = file.path(dirname(fileToRun), paste("FF_", basename(fileToRun), sep = "")),
  sb = "fastq_filter.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  fastqFormat = "sanger",
  indexPosition = NULL,
  qsFilter = NULL,
  maxN = NULL,
  outputFormat = "fastq",
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastqFilter_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastqFilter_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{sb}{Path to fastx_barcode_splitter.pl from FastX toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{fastqFormat}{The fastq format used.}

\item{indexPosition}{Position of index in read}

\item{qsFilter}{Quality score filter}

\item{maxN}{Maximum number of bases below qsFilter}

\item{outputFormat}{Output format (fastq | fasta)}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's fastq_filter
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
