% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_joinWrapper}
\alias{ctk_joinWrapper}
\title{Wrapper function for galaxy's joinWrapper}
\usage{
ctk_joinWrapper(
  file1,
  file2,
  field1,
  field2,
  mode,
  outFile = file.path(dirname(file1), paste0("Unique_", basename(file1))),
  sb = "joinWrapper.py",
  python = "python",
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file1}{File 1 to join.}

\item{file2}{File 2 to join.}

\item{field1}{Field/column for file 1 join.}

\item{field2}{Field/column for file 2 join.}

\item{mode}{Join mode.}

\item{outFile}{Output file}

\item{sb}{Path to tag2collapse.pl from FastX toolkit}

\item{python}{Path to python}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for galaxy's joinWrapper
}
\examples{
\dontrun{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
mutationFile <- system.file("extdata/Fox3_Std_small_mutation.txt",package="clipR")
parsedAlignment <- ctk_parseAlignment(bam,mutationFile=mutationFile)
uniqueTags <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)
ctk_joinWrapper(mutationFile,uniqueTags,4,4,"N",verbose=TRUE)
}
}
\author{
Kathryn Rozen-Gagnon
}
