% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_parseAlignment}
\alias{ctk_parseAlignment}
\title{Wrapper function for ctk's parseAlignment}
\usage{
ctk_parseAlignment(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".bed"),
  sb = "parseAlignment.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  mutationFile = NULL,
  mapQual = NULL,
  minLen = NULL,
  indelToEnd = 5,
  splitDel = FALSE,
  indelInScore = FALSE,
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_parseAlignment_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_parseAlignment_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (SAM).}

\item{outFile}{path to output file (BED).}

\item{sb}{path to parseAlignment.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{mutationFile}{mutation file path, default is NULL.}

\item{mapQual}{minimum map quality, default is NULL.}

\item{minLen}{minimum length of read, default is NULL.}

\item{indelToEnd}{minimum distance from indel to end of read, default is 5.}

\item{splitDel}{whether to split reads with deletions, TRUE or FALSE (default).}

\item{indelInScore}{include indels in mutation score count, TRUE or FALSE (default).}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to BED file.
}
\description{
Wrapper function for ctk's parseAlignment
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta, overwrite=TRUE)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20",verbose = TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- bowtie_align(FqFile_QFColStripped,myIndex, overwrite=TRUE, inputFormat="fastq")
parsedAlignment <- ctk_parseAlignment(bam)
}
\author{
Kathryn Rozen-Gagnon
}
