% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_tag2collapse}
\alias{ctk_tag2collapse}
\title{Wrapper function for ctk's tag2collapse}
\usage{
ctk_tag2collapse(
  filesToRun,
  outFile = file.path(dirname(filesToRun), paste0("TC_", basename(filesToRun))),
  sb = "tag2collapse.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  keepMaxScore = TRUE,
  keepTagName = TRUE,
  weight = TRUE,
  bigFile = TRUE,
  weightInName = TRUE,
  randomBarcode = TRUE,
  seqErrorModel = NULL,
  outputSeqError = NULL,
  em = NULL,
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{sb}{Path to tag2collapse.pl from FastX toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{keepMaxScore}{keep the tag with the most weight (instead of the longest one) as representative}

\item{keepTagName}{do not change tag name (no extra information)}

\item{weight}{consider the weight of each tag}

\item{bigFile}{Set to TRUE when files are big.}

\item{weightInName}{find weight in name}

\item{randomBarcode}{random barcode exists, no collapse for different barcodes}

\item{seqErrorModel}{sequencing error model to use (alignment or em-local or em-global or fix=0.01)}

\item{outputSeqError}{output sequencing errors estimated by the EM algorithm}

\item{em}{EM threshold to infer reliability of each collapsed read (when have random linker, -1=no EM)}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's tag2collapse
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
parsedAlignment <- ctk_parseAlignment(bam)
ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,weightInName=FALSE,verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
