% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_tag2peak}
\alias{ctk_tag2peak}
\title{Wrapper function for ctk's bed2rgb}
\usage{
ctk_tag2peak(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".peak.bed"),
  outBoundary = paste0(file_path_sans_ext(filesToRun), ".boundary.bed"),
  outHalfPH = paste0(file_path_sans_ext(filesToRun), ".halfPF.bed"),
  sb = "tag2peak.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  ss = TRUE,
  valleySeeking = TRUE,
  valleyDepth = 0.9,
  genes = NULL,
  multiTest = FALSE,
  useExpr = FALSE,
  skipOutOfRangePeaks = FALSE,
  pCutOff = 0.01,
  minPH = 2,
  maxPH = -1,
  gap = -1,
  peakPrefix = "Peak",
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2peak_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2peak_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{outFile}{path to output file (BED).}

\item{outBoundary}{output cluster boundaries.}

\item{outHalfPH}{output half peak height boundaries.}

\item{sb}{path to tag2peak.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{bigFile}{big input file, TRUE or FALSE (default).}

\item{ss}{separate the two strands, TRUE (default) or FALSE.}

\item{valleySeeking}{find candidate peaks by valley seeking.}

\item{valleyDepth}{depth of valley if valley seeking (between 0.5 and 1, default is 0.9).}

\item{genes}{custom gene bed file for scan statistics (will override --dbkey)}

\item{multiTest}{do Bonferroni multiple test correction.}

\item{useExpr}{use expression levels given in the score column in the gene bed file for normalization.}

\item{skipOutOfRangePeaks}{Remove out of bounds ranges.}

\item{pCutOff}{threshold of p-value to call peak (e.g. 0.01).}

\item{minPH}{min peak height.}

\item{maxPH}{max peak height.}

\item{gap}{merge cluster peaks closer than the gap (-1, no merge if < 0).}

\item{peakPrefix}{prefix of peak id (Peak) (so output file will look like Peak1, Peak2, etc).}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
path to BED file.
}
\description{
Wrapper function for ctk's bed2rgb
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <-suppressWarnings(bowtie2_index(testFasta, overwrite = TRUE))
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- suppressWarnings(bowtie_align(FqFile_QFColStripped,myIndex, 
overwrite=TRUE, inputFormat="fastq"))
parsedAlignment <- ctk_parseAlignment(bam)
myCollaped <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)
myrgbBed <- ctk_bed2rgb(myCollaped,col="128,0,0")
ctk_tag2profile(myrgbBed,verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
