% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{decompress}
\alias{decompress}
\title{Decompress}
\usage{
decompress(
  fileToDecompress,
  outDir = file_path_sans_ext(fileToDecompress),
  keep = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{fileToDecompress}{File to bzip}

\item{outDir}{Output directory}

\item{keep}{keep (don't delete) input files}

\item{overwrite}{overwrite existing output files}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for bzip2
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
decompress(testFQ,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
