% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastq_quality_filter}
\alias{fastq_quality_filter}
\title{Wrapper function for fastq_quality_filter}
\usage{
fastq_quality_filter(
  fileTofqf,
  outFile = file.path(dirname(fileTofqf), paste0("QF_", basename(fileTofqf))),
  fqf = "fastq_quality_filter",
  qEncoding = 33,
  minimumQuality = 20,
  minimumPercentOfRead = 80,
  stderr = file.path(dirname(fileTofqf), paste0(basename(fileTofqf),
    "_fastq_quality_filter_stderr.txt")),
  stdout = file.path(dirname(fileTofqf), paste0(basename(fileTofqf),
    "_fastq_quality_filter_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{fileTofqf}{path to file to process (fastq).}

\item{outFile}{output file name.}

\item{fqf}{path to fastq_quality_filter from FastX toolkit.}

\item{qEncoding}{quality encoding, set to either 33 (Sanger Phred+33 encoding/Illumina fastq; default) or NULL (Phred+64 fastq).}

\item{minimumQuality}{minimum quality score to keep (default is 20).}

\item{minimumPercentOfRead}{minimum percent of bases that must have [-q] quality (default is 80).}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastq_quality_filter
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_filter(FqFile)
}
\author{
Kathryn Rozen-Gagnon
}
