% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastq_quality_trimmer}
\alias{fastq_quality_trimmer}
\title{Wrapper function for fastq_quality_trimmer}
\usage{
fastq_quality_trimmer(
  fileTofqf,
  outFile = file.path(dirname(fileTofqf), paste0("QF_", basename(fileTofqf))),
  fqf = "fastq_quality_trimmer",
  qualityThreshold = 5,
  minimumLength = 20,
  stderr = file.path(dirname(fileTofqf), paste0(basename(fileTofqf),
    "_fastq_quality_trimmer_stderr.txt")),
  stdout = file.path(dirname(fileTofqf), paste0(basename(fileTofqf),
    "_fastq_quality_trimmer_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{fileTofqf}{File to process.}

\item{outFile}{Output file name.}

\item{fqf}{Path to fastq_quality_trimmer from FastX toolkit}

\item{qualityThreshold}{Minimum quality score to keep.}

\item{minimumLength}{Minimum percent of bases that must have [-q] quality.}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastq_quality_trimmer
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
}
\author{
Kathryn Rozen-Gagnon
}
