% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastq_to_fasta}
\alias{fastq_to_fasta}
\alias{fastx_qtoa}
\title{Wrapper function for fastq_to_fasta}
\usage{
fastx_qtoa(
  fileToc,
  fqa = "fastq_to_fasta",
  discard_N = T,
  rename = F,
  outFile = paste0(file_path_sans_ext(fileToc), ".fa"),
  writelog = T,
  qEncoding = 33,
  stderr = file.path(dirname(fileToc), paste0(basename(fileToc), "_qtoa_stderr.txt")),
  stdout = file.path(dirname(fileToc), paste0(basename(fileToc), "_qtoa_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileToc}{path to file to process (fastq).}

\item{fqa}{path to fastx_trimmer from FastX toolkit.}

\item{discard_N}{discard reads with unknown nucleotides (N), TRUE (default)  or FALSE.}

\item{rename}{replace sequence names with number identifiers, TRUE or FALSE (default).}

\item{outFile}{output file path.}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}

\item{qEncoding}{quality encoding, set to either 33 (Sanger Phred+33 encoding/Illumina fastq; default) or NULL (Phred+64 fastq).}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to trimmed file
}
\description{
Wrapper function for fastq_to_fasta
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_fa <- fastx_qtoa(FqFile)
}
\author{
Kathryn Rozen-Gagnon
}
