% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_barcode_splitter}
\alias{fastx_barcode_splitter}
\title{Wrapper function for fastx_barcode_splitter}
\usage{
fastx_barcode_splitter(
  fileTofxc,
  bcFile,
  mismatches = 0,
  fbs = "fastx_barcode_splitter.pl",
  stderr = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_barcode_splitter_stderr.txt")),
  stdout = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_barcode_splitter_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofxc}{File to process.}

\item{bcFile}{Barcode file}

\item{mismatches}{Number of mismatches allowed.}

\item{fbs}{Path to fastx_barcode_splitter.pl from FASTX toolkit}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}
}
\value{
output3, path to split files
}
\description{
Wrapper function for fastx_barcode_splitter
}
\author{
Kathryn Rozen-Gagnon
}
