% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_clipper}
\alias{fastx_clipper}
\title{Wrapper function for fastx_clipper}
\usage{
fastx_clipper(
  fileTofqs,
  outFile = paste0(file_path_sans_ext(fileTofqs), "_clip.", file_ext(fileTofqs)),
  fqc = "fastx_clipper",
  length = 18,
  adaptor = "GTGTCAGTCACTTCCAGCGG",
  writelog = T,
  stderr = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_clipper_stderr.txt")),
  stdout = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_clipper_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqs}{File to process.}

\item{outFile}{Output file path}

\item{fqc}{Path to fastx_clipper from FastX toolkit}

\item{length}{Length for fastx_clipper}

\item{adaptor}{Adapter to remove}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}
}
\value{
Path to clipped file
}
\description{
Wrapper function for fastx_clipper
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
}
\author{
Kathryn Rozen-Gagnon
}
