% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_collapser}
\alias{fastx_collapser}
\title{Wrapper function for fastx_collapser}
\usage{
fastx_collapser(
  fileTofxc,
  outFile = file.path(dirname(fileTofxc), gsub("\\\\.fastq|\\\\.fq|\\\\.fa|\\\\.fasta",
    "_collapse.fasta", basename(fileTofxc))),
  fxc = "fastx_collapser",
  qEncoding = NULL,
  stderr = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_collapse_stderr.txt")),
  stdout = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_collapse_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{fileTofxc}{path to file to process (fastq or fasta).}

\item{outFile}{output file name (will be formatted as a fasta).}

\item{fxc}{path to fastx_collapser from FastX toolkit.}

\item{qEncoding}{quality encoding, set to either 33 (Sanger Phred+33 encoding/Illumina fastq; default) or NULL (Phred+64 fastq or fasta)}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to unzipped file
}
\description{
Wrapper function for fastx_collapser
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
}
\author{
Kathryn Rozen-Gagnon
}
