% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_quality_stats}
\alias{fastx_quality_stats}
\title{Wrapper function for fastx_quality_stats}
\usage{
fastx_quality_stats(
  fileTofqs,
  outFile = file.path(dirname(fileTofqs),
    gsub("\\\\.fastq|\\\\.fq|fq\\\\.gz|fastq\\\\.gz|\\\\.fasta|\\\\.fa|fa\\\\.gz|fasta\\\\.gz",
    ".txt", basename(fileTofqs))),
  fqs = "fastx_quality_stats",
  qEncoding = 33,
  stderr = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_quality_stats_stderr.txt")),
  stdout = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_quality_stats_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqs}{path to file to process (fastq or fasta - fasta will only return nucleotide distribution).}

\item{outFile}{output file name.}

\item{fqs}{path to fastx_quality_stats from FastX toolkit.}

\item{qEncoding}{quality encoding,  set to either 33 (Sanger Phred+33 encoding/Illumina fastq; default) or NULL (Phred+64 fastq or fasta).}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to unzipped file.
}
\description{
Wrapper function for fastx_quality_stats
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
Fq_Stats <- fastx_quality_stats(FqFile_QF)
}
\author{
Kathryn Rozen-Gagnon
}
