% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_trimmer}
\alias{fastx_trimmer}
\title{Wrapper function for fastx_trimmer}
\usage{
fastx_trimmer(
  fileTofqt,
  fqt = "fastx_trimmer",
  read_start = 10,
  read_end = NULL,
  qEncoding = NULL,
  outFile = paste0(file_path_sans_ext(fileTofqt), "_trim.", file_ext(fileTofqt)),
  writelog = T,
  stderr = file.path(dirname(fileTofqt), paste0(basename(fileTofqt),
    "_trimmer_stats_stderr.txt")),
  stdout = file.path(dirname(fileTofqt), paste0(basename(fileTofqt),
    "_trimmer_stats_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqt}{path to file to process (fastq or fasta).}

\item{fqt}{path to fastx_trimmer from FastX toolkit.}

\item{read_start}{read starting base (default is 10).}

\item{read_end}{read ending base (default is NULL, set to integer to trim).}

\item{qEncoding}{quality encoding, set to either 33 (Sanger Phred+33 encoding/Illumina fastq) or NULL (Phred+64 fastq or fasta; default).}

\item{outFile}{output file path.}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to trimmed file.
}
\description{
Wrapper function for fastx_trimmer
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_trimmed <- fastx_trimmer(FqFile,read_start = 5, read_end = 50)
}
\author{
Kathryn Rozen-Gagnon
}
