% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{fetchSequencesForCLIP}
\alias{fetchSequencesForCLIP}
\title{Retrieve sequences from under peaks/regions}
\usage{
fetchSequencesForCLIP(
  peaks,
  resize = NULL,
  fasta,
  add5 = 0,
  add3 = 0,
  verbose = FALSE,
  bedHeader = TRUE,
  bedSep = "\\t"
)
}
\arguments{
\item{peaks}{CLIP peaks to process; accepts path to peak files (BED file or BED-formatted tab-delimited text files) or R objects (GRanges or BED-formatted data.frame); peak locations must be unique.}

\item{resize}{size of window in bp for resizing around peak center, default is NULL (set by specifying integer).}

\item{fasta}{path to genome file (fasta).}

\item{add5}{bp to add to 5' of resized peak, default is 0 (set by specifying integer).}

\item{add3}{bp to add to 3' of resized peak, default is 0 (set by specifying integer).}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{bedHeader}{if peak file contains column headers, TRUE (default) or FALSE; if TRUE, column names must be "seqnames" (chromosome), "start" (peak start), "end" (peak end), "strand" (peak strand).}

\item{bedSep}{separator in BED file.}
}
\value{
path to unzipped file.
}
\description{
Retrieve sequences from under peaks/regions
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
decom <- decompress(testFQ,overwrite=TRUE)
com <- compress(decom,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
