% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{annotatePeaksWithPatterns}
\alias{annotatePeaksWithPatterns}
\title{annotatePeaksWithPatterns}
\usage{
annotatePeaksWithPatterns(
  peaks,
  fasta,
  patterns,
  resize = 64,
  add5 = 0,
  add3 = 0,
  verbose = FALSE,
  checkReverse = TRUE,
  bedHeader = TRUE,
  bedSep = "\\t"
)
}
\arguments{
\item{peaks}{CLIP peaks to process}

\item{fasta}{Genome fasta file}

\item{patterns}{Patterns to scan in CLIP peaks}

\item{resize}{Size of window in bp for resizing around peak center}

\item{add5}{Bp to add to 5' of resized peak}

\item{add3}{Bp to add to 3' of resized peak}

\item{verbose}{Print messages, TRUE (default) or FALSE}

\item{checkReverse}{Check the reverse strand for pattern.}

\item{bedHeader}{TRUE if peak file contains column headers.}

\item{bedSep}{Seperator in BED file.}
}
\value{
Path
}
\description{
Search for small RNA targets, or any pattern, in peaks
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
}
\author{
Kathryn Rozen-Gagnon
}
