% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{annotatePeaksWithPatterns}
\alias{annotatePeaksWithPatterns}
\title{Search for small RNA targets, or any pattern, in peaks}
\usage{
annotatePeaksWithPatterns(
  peaks,
  fasta,
  patterns,
  resize = 64,
  add5 = 0,
  add3 = 0,
  verbose = FALSE,
  checkReverse = TRUE,
  bedHeader = TRUE,
  bedSep = "\\t"
)
}
\arguments{
\item{peaks}{CLIP peaks to process; accepts path to peak files (BED file or BED-formatted tab-delimited text files) or R objects (GRanges or BED-formatted data.frame); peak locations must be unique.}

\item{fasta}{path to genome file (fasta).}

\item{patterns}{patterns to scan in CLIP peaks (character vector, DNAStringSet, or file path to a fasta sequence).}

\item{resize}{size of window in bp for resizing around peak center.}

\item{add5}{bp to add to 5' of resized peak, default is 0 (set by specifying integer).}

\item{add3}{bp to add to 3' of resized peak, default is 0 (set by specifying integer).}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{checkReverse}{check the reverse strand for pattern.}

\item{bedHeader}{if peak file contains column headers, TRUE (default) or FALSE; if TRUE, column names must be "seqnames" (chromosome), "start" (peak start), "end" (peak end), "strand" (peak strand).}

\item{bedSep}{separator in BED file.}
}
\value{
Path
}
\description{
Search for small RNA targets, or any pattern, in peaks
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
}
\author{
Kathryn Rozen-Gagnon
}
