% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{bamtobed}
\alias{bamtobed}
\title{Convert Bam to Bed}
\usage{
bamtobed(file, outFile = gsub("\\\\.bam", ".bed", file), filtDup = FALSE)
}
\arguments{
\item{file}{path to file to process (BAM).}

\item{outFile}{path to output file (BED).}

\item{filtDup}{output index name.}
}
\value{
path to BED.
}
\description{
Convert Bam to Bed
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- suppressWarnings(bowtie2_index(testFasta, overwrite = TRUE))
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_QFCollapsed)
FqFile_QFColStpClipped <- fastx_clipper(FqFile_QFColStripped)
bam <- suppressWarnings(bowtie_align(FqFile_QFColStpClipped,myIndex, overwrite = TRUE))
bamtobed(bam)
}
\author{
Kathryn Rozen-Gagnon
}
