% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{bowtie2_index}
\alias{bowtie2_index}
\title{bowtie2_index}
\usage{
bowtie2_index(
  genomeFasta,
  outIndex = gsub("\\\\.fa", "", genomeFasta),
  overwrite = TRUE,
  threads = 1
)
}
\arguments{
\item{genomeFasta}{File to process.}

\item{outIndex}{Output index name}

\item{overwrite}{Overwrite if directory exists.}

\item{threads}{Number of threads to use.}
}
\value{
Path to index
}
\description{
Make index for Rbowtie2
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
bowtie2_index(testFasta)
}
\author{
Kathryn Rozen-Gagnon
}
