% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{bowtie2_index}
\alias{bowtie2_index}
\title{bowtie2_index}
\usage{
bowtie2_index(
  genomeFasta,
  outIndex = gsub("\\\\.fa", "", genomeFasta),
  overwrite = F,
  threads = 1
)
}
\arguments{
\item{genomeFasta}{path to file to process.}

\item{outIndex}{path to output index.}

\item{overwrite}{overwrite existing output files, TRUE or FALSE (default).}

\item{threads}{number of threads to use, default is 1.}
}
\value{
Path to index
}
\description{
Make index for Rbowtie2
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
bowtie2_index(testFasta, overwrite = TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
