% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{bowtie_align}
\alias{bowtie_align}
\title{Alignment using Rbowtie2}
\usage{
bowtie_align(
  fq,
  index,
  bam = file.path(dirname(fq), paste0("Sorted_", basename(fq), ".bam")),
  format = "fasta",
  maxMismatches = 1,
  seedSubString = 18,
  threads = 1,
  report_k = NULL
)
}
\arguments{
\item{fq}{File to process.}

\item{index}{Index name}

\item{bam}{Output bam name}

\item{format}{Format of reads (fastq or fasta)}

\item{maxMismatches}{max mismatches}

\item{seedSubString}{length of seed substrings}

\item{threads}{Number of threads to use}

\item{report_k}{number of alignments to report}
}
\value{
Path to BAM
}
\description{
Alignment using Rbowtie2
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bowtie_align(FqFile_QFColStripped,myIndex)
}
\author{
Kathryn Rozen-Gagnon
}
