% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{bzip2}
\alias{bzip2}
\title{Wrapper function for bzip2}
\usage{
bzip2(
  fileToBzip2,
  bzip2 = "Bzip2",
  keep = TRUE,
  force = FALSE,
  small = FALSE,
  blockSize = 1,
  stderr = paste0(getwd(), "gunzip_stderr"),
  stdout = paste0(getwd(), "gunzip_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileToBzip2}{File to bzip.}

\item{bzip2}{Path to bzip2.}

\item{keep}{keep (don't delete) input files.}

\item{force}{overwrite existing output files.}

\item{small}{use less memory (at most 2500k).}

\item{blockSize}{set block size to 100k .. 900k.}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs.}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file.
}
\description{
Wrapper function for bzip2
}
\author{
Kathryn Rozen-Gagnon
}
