% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{chimera_Process}
\alias{chimera_Process}
\title{chimera_Process}
\usage{
chimera_Process(
  bams,
  knownMiRNAs,
  genomeIndex,
  exclude,
  bpparam = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bams}{BAM files mapped to the genome to process}

\item{knownMiRNAs}{FASTA file containing annotated miRNA or other small RNA sequences}

\item{genomeIndex}{Full genome index}

\item{exclude}{Names of miRNAs or small RNA sequences to remove, character vector}

\item{bpparam}{TRUE or FALSE (default)}

\item{verbose}{print messages, TRUE (default) or FALSE}
}
\value{
Path
}
\description{
Chimera process CLIP data
}
\examples{
testFastq <- system.file("extdata/SRR1742056.fastq.gz",package="CLIPflexR")
testMiRNA <- system.file("extdata/hsa_mature.fa",package="CLIPflexR")
testMiRNA <- system.file("extdata/hsa_mature.fa",package="CLIPflexR")
}
\author{
Kathryn Rozen-Gagnon
}
