% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chimera_Functions.R
\docType{methods}
\name{chimera_process}
\alias{chimera_process}
\alias{chimera_Process}
\title{Merge read sequence with bedfile}
\usage{
chimera_Process(
  bams,
  knownMiRNAs,
  genomeIndex,
  exclude,
  bpparam = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bams}{BAM files to process.}

\item{knownMiRNAs}{knownMiRNAs FASTA file.}

\item{genomeIndex}{Full genome index}

\item{exclude}{contaminating sequences to remove}
}
\value{
Path
}
\description{
Merge read sequence with bedfile
}
\author{
Kathryn Rozen-Gagnon
}
