% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_align.R
\docType{methods}
\name{ClIP_align}
\alias{ClIP_align}
\title{Align to Genome}
\usage{
ClIP_align(
  samID,
  res_dir = NULL,
  genome_idx = NULL,
  aligner = NULL,
  samSheet = NULL,
  thread = 4,
  bwa = NULL
)
}
\arguments{
\item{samID}{sample id in the sample sheet}

\item{res_dir}{result directory}

\item{genome_idx}{file path of indexed genome}

\item{aligner}{Assign aligner bowtie2/bwa_mem/bwa_alb/subread/subjunc/hisat2}

\item{samSheet}{sample sheet}

\item{thread}{thread used in the processing, default is 4}

\item{bwa}{exercutable file path of bwa}
}
\value{
sorted bam file
}
\description{
Align to Genome
}
