% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{CLIP_bw2}
\alias{CLIP_bw2}
\title{Build bigwigs}
\usage{
CLIP_bw2(sort_bam, res_dir = NULL, normalized = TRUE)
}
\arguments{
\item{sort_bam}{sorted bam file}

\item{res_dir}{result directory}

\item{normalized}{Normalised to total reads}
}
\value{
bigwig
}
\description{
Build bigwigs
}
