% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_bw2.R
\docType{methods}
\name{ClIP_bw2}
\alias{ClIP_bw2}
\title{Build bigwigs}
\usage{
ClIP_bw2(sort_bam, res_dir = NULL, normalized = TRUE)
}
\arguments{
\item{sort_bam}{sorted bam file}

\item{res_dir}{result directory}

\item{normalized}{Normalised to total reads.}
}
\value{
bigwig
}
\description{
Build bigwigs
}
