% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{compress}
\alias{compress}
\title{Wrapper function for Compress}
\usage{
compress(
  fileToCompress,
  keep = TRUE,
  methodCompress = "gunzip",
  overwrite = FALSE
)
}
\arguments{
\item{fileToCompress}{File to bzip.}

\item{keep}{keep (don't delete) input files.}

\item{methodCompress}{Method to use for compression}

\item{overwrite}{overwrite existing output files.}
}
\value{
Path to unzipped file.
}
\description{
Wrapper function for Compress
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
decom <- decompress(testFQ,overwrite=TRUE)
com <- compress(decom,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
