% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{countFromBed}
\alias{countFromBed}
\title{countFromBed to make count matrix}
\usage{
countFromBed(Bed, GR, notStranded = TRUE, interFeature = FALSE)
}
\arguments{
\item{Bed}{BED bed files containing mapped reads}

\item{GR}{BED file containing genomic ranges over which to count}

\item{notStranded}{Stranded or not stranded TRUE (default) or FALSE}

\item{interFeature}{Count reads mapping to multiple features TRUE or FALSE (default)}
}
\value{
counting matrix
}
\description{
count from mapped beds to make count matrix
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
bowtie2_index(testFasta)
}
\author{
Kathryn Rozen-Gagnon
}
