% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFromBed.R
\docType{methods}
\name{countFromBed}
\alias{countFromBed}
\title{make count matrix from bed}
\usage{
countFromBed(Bed, GR, notStranded = TRUE, interFeature = FALSE)
}
\arguments{
\item{GR}{count reads over these genomic ranges.}

\item{notStranded}{Stranded or not stranded TRUE or FALSE.}

\item{interFeature}{Count reads mapping to multiple features TRUE or FALSE.}

\item{bed}{mapped reads, bed files.}
}
\value{
counting matrix
}
\description{
make count matrix from bed
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
bowtie2_index(testFasta)
}
\author{
Kathryn Rozen-Gagnon
}
