% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{countFromBed}
\alias{countFromBed}
\title{countFromBed to make count matrix}
\usage{
countFromBed(Bed, GR, notStranded = TRUE, interFeature = FALSE)
}
\arguments{
\item{Bed}{BED bed files containing mapped reads.}

\item{GR}{genomic ranges object over which to count.}

\item{notStranded}{if strand information should be considered when counting, TRUE (ignore strand) or FALSE (consider strand; default).}

\item{interFeature}{discard reads mapping to multiple features, TRUE or FALSE (default).}
}
\value{
count matrix.
}
\description{
count from mapped beds to make count matrix
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
bowtie2_index(testFasta, overwrite = TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
