% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFromBed.R
\docType{methods}
\name{countFromBed}
\alias{countFromBed}
\title{make count matrix from bed}
\usage{
countFromBed(Bed, GR, notStranded = TRUE, interFeature = FALSE)
}
\arguments{
\item{GR}{GR need add description}

\item{notStranded}{Stranded or not stranded}

\item{interFeature}{interfeature # need add description}

\item{bed}{Bed file}
}
\value{
counting matrix
}
\description{
make count matrix from bed
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
bowtie2_index(testFasta)
}
\author{
Kathryn Rozen-Gagnon
}
