% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_bed2rgb}
\alias{ctk_bed2rgb}
\title{Wrapper function for ctk's bed2rgb}
\usage{
ctk_bed2rgb(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".RGB.", file_ext(filesToRun)),
  sb = "bed2rgb.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  col = "blue",
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_bed2rgb_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_bed2rgb_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{outFile}{path to output file (BED).}

\item{sb}{path to bed2rgb.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{col}{color to include in BED rgb column (default is "blue").}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to BED file with color specified.
}
\description{
Wrapper function for ctk's bed2rgb
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package ="CLIPflexR")
myIndex <-suppressWarnings(bowtie2_index(testFasta, overwrite = TRUE))
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package ="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20")
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- suppressWarnings(bowtie_align(FqFile_QFColStripped,myIndex, overwrite=TRUE, inputFormat="fastq"))
parsedAlignment <- ctk_parseAlignment(bam)
myCollapsed <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)
ctk_bed2rgb(myCollapsed,col="128,0,0")
}
\author{
Kathryn Rozen-Gagnon
}
