% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_cims}
\alias{ctk_cims}
\title{Wrapper function for ctk's CIMS}
\usage{
ctk_cims(
  filesToRun,
  mutationBedFile,
  outFile = paste(file_path_sans_ext(fileToRun), "CIMS", "txt", sep = "."),
  sb = "CIMS.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  mutationSize = 1,
  permutations = 5,
  trackMutationPos = FALSE,
  noSparseCorrect = FALSE,
  FDR = 1,
  mfr = 0,
  cacheDir = paste0(filesToRun, "_cache", paste(sample(letters, 10), collapse = "")),
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun), "_cims_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun), "_cims_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{mutationBedFile}{mutation BED file.}

\item{outFile}{path to output file (BED & text).}

\item{sb}{path to CIMS.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{bigFile}{big file, TRUE or FALSE (default).}

\item{mutationSize}{mutation size (default is 1).}

\item{permutations}{number of iterations for permutation (default is 5).}

\item{trackMutationPos}{track mutation position relative to read start, TRUE or FALSE (default).}

\item{noSparseCorrect}{no sparcity correction, TRUE or FALSE (default).}

\item{FDR}{threshold of FDR (default is 1).}

\item{mfr}{threshold of m-over-k-ratio (default is 0).}

\item{cacheDir}{name for cache directory.}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to CIMS text & bed files.
}
\description{
Wrapper function for ctk's CIMS
}
\examples{
\dontrun{
mutations <- system.file("extdata/BrdU.Fox.pool.tag.uniq.mutation.small.txt",package="CLIPflexR")
delBed <- ctk_getMutationType(mutations)
ctk_cims("~/Downloads/uniq_tags_mutations/Fox.pool.tag.uniq.rgb.bed",delBed,verbose=TRUE)
}
}
\author{
Kathryn Rozen-Gagnon
}
