% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_cits}
\alias{ctk_cits}
\title{Wrapper function for ctk's CITS}
\usage{
ctk_cits(
  filesToRun,
  outFile = paste(file_path_sans_ext(fileToRun), "CITS", "bed", sep = "."),
  sb = "CITS.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  pCutOff = 0.01,
  multiTest = TRUE,
  gap = 25,
  cacheDir = paste0(filesToRun, "_cache", paste(sample(letters, 10), collapse = "")),
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{sb}{Path to fastx_barcode_splitter.pl from FastX toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{bigFile}{big file}

\item{pCutOff}{p-value cut off}

\item{multiTest}{perform multiple testing correction.}

\item{gap}{gap size}

\item{cacheDir}{Name for cache directory.}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's CITS
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_QFCollapsed)
}
\author{
Kathryn Rozen-Gagnon
}
