% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_fastq2collapse}
\alias{ctk_fastq2collapse}
\title{Wrapper function for ctk's fastq2collapse}
\usage{
ctk_fastq2collapse(
  filesToRun,
  outFile = file.path(dirname(fileToRun), paste("Collapsed_", basename(fileToRun), sep =
    "")),
  sb = "fastq2collapse.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastq2collapse_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastq2collapse_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (fastq).}

\item{outFile}{output file (fastq).}

\item{sb}{path to fastq2collapse.pl from CTK.}

\item{perl}{path to PERL}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
Path to collapsed fastq file.
}
\description{
Wrapper function for ctk's fastq2collapse
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
