% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_fastqFilter}
\alias{ctk_fastqFilter}
\title{Wrapper function for ctk's fastq_filter}
\usage{
ctk_fastqFilter(
  filesToRun,
  outFile = file.path(dirname(fileToRun), paste("FF_", basename(fileToRun), sep = "")),
  sb = "fastq_filter.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  fastqFormat = "sanger",
  indexPosition = NULL,
  qsFilter = NULL,
  maxN = NULL,
  outputFormat = "fastq",
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastqFilter_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_fastqFilter_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process  (fastq).}

\item{outFile}{output file (fastq).}

\item{sb}{path to fastq_filter.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{fastqFormat}{fastq format used, can be "sanger" (default) or "solexa".}

\item{indexPosition}{position and sequence of index in read, default is NULL; set "position:seqeunce" to specify (e.g. "1:CATCGC").}

\item{qsFilter}{set quality score filter, default is NULL; set method:start-end:score to specify (e.g. "mean:0-29:20"; starts/ends are 0-based).}

\item{maxN}{maximum number of unknown nucleotides (N) allowed, default is NULL; set to integer to specify.}

\item{outputFormat}{output format, "fastq" (default) or "fasta".}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
Path to filtered file in specified format.
}
\description{
Wrapper function for ctk's fastq_filter
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
