% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_getMutationType}
\alias{ctk_getMutationType}
\title{Wrapper function for ctk's getMutationType}
\usage{
ctk_getMutationType(
  filesToRun,
  outFile = paste(file_path_sans_ext(fileToRun), mutationType, "bed", sep = "."),
  sb = "getMutationType.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  mutationType = "del",
  summaryStat = FALSE,
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{sb}{Path to getMutationType.pl from CTK toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{mutationType}{mutation size}

\item{summaryStat}{Create summary file}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's getMutationType
}
\examples{
mutations <- system.file("extdata/BrdU.Fox.pool.tag.uniq.mutation.small.txt",package="CLIPflexR")
ctk_getMutationType(mutations)
}
\author{
Kathryn Rozen-Gagnon
}
