% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_getMutationType}
\alias{ctk_getMutationType}
\title{Wrapper function for ctk's getMutationType}
\usage{
ctk_getMutationType(
  filesToRun,
  outFile = paste(file_path_sans_ext(fileToRun), mutationType, "bed", sep = "."),
  sb = "getMutationType.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  mutationType = "del",
  summaryStat = FALSE,
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_getMutationType_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_getMutationType_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{outFile}{output file (BED).}

\item{sb}{path to ctk_getMutationType.pl from CTK toolkit.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{mutationType}{mutation type, "del" (deletions; default),"ins" (insertions), "sub" (substitutions).}

\item{summaryStat}{create summary file, TRUE or FALSE (default).}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to getMutationType bed file.
}
\description{
Wrapper function for ctk's getMutationType
}
\examples{
mutations <- system.file("extdata/BrdU.Fox.pool.tag.uniq.mutation.small.txt",package="CLIPflexR")
ctk_getMutationType(mutations)
}
\author{
Kathryn Rozen-Gagnon
}
