% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_joinWrapper}
\alias{ctk_joinWrapper}
\title{Wrapper function for galaxy's joinWrapper}
\usage{
ctk_joinWrapper(
  file1,
  file2,
  field1,
  field2,
  mode,
  outFile = file.path(dirname(file1), paste0("Unique_", basename(file1))),
  sb = "joinWrapper.py",
  python = "python",
  stderr = file.path(dirname(file1), paste0(basename(file1),
    "_ctk_joinWrapper_stderr.txt")),
  stdout = file.path(dirname(file1), paste0(basename(file1),
    "_ctk_joinWrapper_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{file1}{path to file 1 to join (BED or tab delimited text).}

\item{file2}{path to file 2 to join (BED or tab delimited text).}

\item{field1}{field/column for file 1 join.}

\item{field2}{field/column for file 2 join.}

\item{mode}{join mode.}

\item{outFile}{path to output file.}

\item{sb}{path to joinWrapper.py from Galaxy}

\item{python}{path to python.}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}
}
\value{
path to joined file.
}
\description{
Wrapper function for galaxy's joinWrapper
}
\examples{
\dontrun{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta, overwrite=TRUE)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- bowtie_align(FqFile_QFColStripped,myIndex, overwrite=TRUE, inputFormat="fastq")
mutationFile <- system.file("extdata/Fox3_Std_small_mutation.txt",package="CLIPflexR")
parsedAlignment <- ctk_parseAlignment(bam,mutationFile=mutationFile)
uniqueTags <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)
ctk_joinWrapper(mutationFile,uniqueTags,4,4,"N",verbose=TRUE)
}
}
\author{
Kathryn Rozen-Gagnon
}
