% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_stripBarcode}
\alias{ctk_stripBarcode}
\title{Wrapper function for ctk's stripBarcode}
\usage{
ctk_stripBarcode(
  filesToRun,
  outFile = paste(file_path_sans_ext(fileToRun), "_rm5.", file_ext(fileToRun), sep =
    ""),
  sb = "stripBarcode.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  linkerlength = 27,
  inputFormat = "fasta",
  barcodeStartWith = NULL,
  barcodeEndWith = NULL,
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{sb}{Path to stripBarcode.pl from CTK.}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{linkerlength}{length of barcode/linker sequences.}

\item{inputFormat}{Input file format (fasta|fastq)}

\item{barcodeStartWith}{Filter sequences based on the starting nucleotides in the barcode}

\item{barcodeEndWith}{Filter sequences based on the ending nucleotides in the barcode}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's stripBarcode
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
}
\author{
Kathryn Rozen-Gagnon
}
