% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_tag2collapse}
\alias{ctk_tag2collapse}
\title{Wrapper function for ctk's tag2collapse}
\usage{
ctk_tag2collapse(
  filesToRun,
  outFile = file.path(dirname(filesToRun), paste0("TC_", basename(filesToRun))),
  sb = "tag2collapse.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  keepMaxScore = TRUE,
  keepTagName = TRUE,
  weight = TRUE,
  bigFile = FALSE,
  weightInName = TRUE,
  randomBarcode = TRUE,
  seqErrorModel = TRUE,
  outputSeqError = NULL,
  em = NULL,
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2collapse_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2collapse_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{filesToRun}{path to file to process (BED).}

\item{outFile}{path to output file}

\item{sb}{path to tag2collapse.pl from CTK.}

\item{perl}{path to PERL.}

\item{PATHTOPERLLIB}{path to PERL5LIB.}

\item{keepMaxScore}{keep the tag with the most weight (instead of the longest one) as representative.}

\item{keepTagName}{do not change tag name (no extra information).}

\item{weight}{consider the weight of each tag.}

\item{bigFile}{Set to TRUE when files are big.}

\item{weightInName}{find weight in name.}

\item{randomBarcode}{random barcode exists, no collapse for different barcodes.}

\item{seqErrorModel}{sequencing error model to use (alignment or em-local or em-global or fix=0.01).}

\item{outputSeqError}{output sequencing errors estimated by the EM algorithm.}

\item{em}{EM threshold to infer reliability of each collapsed read (when have random linker, -1=no EM).}

\item{stderr}{path to stdout file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{Additional arguments to be passed to system call.}

\item{verbose}{print messages, TRUE or FALSE (default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
path to collapsed BED file.
}
\description{
Wrapper function for ctk's tag2collapse
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <-suppressWarnings(bowtie2_index(testFasta, overwrite = TRUE))
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- bowtie_align(FqFile_QFColStripped,myIndex, overwrite=TRUE, inputFormat="fastq")
parsedAlignment <- ctk_parseAlignment(bam)
ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName = FALSE,verbose = TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
