% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_ctk.R
\docType{methods}
\name{ctk_tag2profile}
\alias{ctk_tag2profile}
\title{Wrapper function for ctk's bed2rgb}
\usage{
ctk_tag2profile(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".out"),
  outFile2 = NULL,
  sb = "tag2profile.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  weight = FALSE,
  weightAvg = FALSE,
  ss = TRUE,
  exact = TRUE,
  nz = FALSE,
  ext5 = NULL,
  ext3 = NULL,
  chromLen = NULL,
  region = NULL,
  minBlockSize = 2e+06,
  windowSize = 100,
  stepSize = 20,
  outputFormat = "bedgraph",
  normalization = "none",
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filesToRun}{File to process.}

\item{outFile}{Output file}

\item{outFile2}{Output 2 file}

\item{sb}{Path to tag2collapse.pl from FastX toolkit}

\item{perl}{Path to PERL}

\item{PATHTOPERLLIB}{Path to PERL5LIB.}

\item{bigFile}{TRUE when working with a big file}

\item{weight}{weight counts according to the score of each tag}

\item{weightAvg}{weight average the score of each tag}

\item{ss}{separate strand}

\item{exact}{exact count at each nucleotide}

\item{nz}{don't print zeroes (works for sgr and bed)}

\item{ext5}{extension of tags at the 5' end}

\item{ext3}{extension of tags at the 3' end}

\item{chromLen}{chrom length file}

\item{region}{a bed file with regions to count tag numbers. If not specified, count in moving windows}

\item{minBlockSize}{minimum number of lines to read in each block for a big file}

\item{windowSize}{Window size}

\item{stepSize}{Step size}

\item{outputFormat}{output format ([bed] or bedgraph or sgr)}

\item{normalization}{normalization ([none] or rpkm or multiply={1.3})}

\item{stderr}{Path to stdout file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for ctk's bed2rgb
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
parsedAlignment <- ctk_parseAlignment(bam)
myCollaped <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,weightInName=FALSE,verbose=TRUE)
myrgbBed <- ctk_bed2rgb(myCollaped,col="128,0,0")
ctk_tag2profile(myrgbBed,verbose=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
