% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{decompress}
\alias{decompress}
\title{Wrapper function for gunzip, bunzip}
\usage{
decompress(
  fileToDecompress,
  outDir = file_path_sans_ext(fileToDecompress),
  keep = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{fileToDecompress}{path to file to decompress.}

\item{outDir}{output directory.}

\item{keep}{keep (don't delete) input files, TRUE (default) or FALSE.}

\item{overwrite}{overwrite existing output files, TRUE or FALSE (default).}
}
\value{
path to unzipped file.
}
\description{
Wrapper function for gunzip, bunzip
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
decompress(testFQ,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
