% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{extract_unmapped}
\alias{extract_unmapped}
\title{extract_unmapped}
\usage{
extract_unmapped(bam, outfa = NULL)
}
\arguments{
\item{bam}{BAM file}

\item{outfa}{PATH to output FASTA (default is same directory as input BAM)}
}
\value{
Path to FASTA file
}
\description{
Extracts unmapped reads from BAM and writes them to a FASTA file
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
unmapped <- extract_unmapped(bam)
unmapped
}
\author{
Kathryn Rozen-Gagnon
}
