% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastq_quality_filter}
\alias{fastq_quality_filter}
\title{Wrapper function for fastq_quality_filter}
\usage{
fastq_quality_filter(
  fileTofqf,
  outFile = file.path(dirname(fileTofqf), paste0("QF_", basename(fileTofqf))),
  fqf = "fastq_quality_filter",
  qEncoding = 33,
  minimumQuality = 20,
  minimumPercentOfRead = 80,
  stderr = paste0(getwd(), "fastq_quality_filter_stderr"),
  stdout = paste0(getwd(), "fastq_quality_filter_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqf}{File to process.}

\item{outFile}{Output file name.}

\item{fqf}{Path to fastq_quality_filter from FastX toolkit}

\item{qEncoding}{Quality encoding}

\item{minimumQuality}{Minimum quality score to keep.}

\item{minimumPercentOfRead}{Minimum percent of bases that must have [-q] quality.}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastq_quality_filter
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_filter(FqFile)
}
\author{
Kathryn Rozen-Gagnon
}
