% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_barcode_splitter}
\alias{fastx_barcode_splitter}
\title{Wrapper function for fastx_collapser}
\usage{
fastx_barcode_splitter(
  fileTofxc,
  bcFile,
  mismatches = 0,
  fbs = "fastx_barcode_splitter.pl",
  stderr = paste0(getwd(), "fastx_barcode_splitter_stderr"),
  stdout = paste0(getwd(), "fastx_barcode_splitter_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofxc}{File to process.}

\item{bcFile}{Barcode file}

\item{mismatches}{Number of mismatches allowed.}

\item{fbs}{Path to fastx_barcode_splitter.pl from FastX toolkit}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastx_collapser
}
\author{
Kathryn Rozen-Gagnon
}
