% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_barcode_splitter}
\alias{fastx_barcode_splitter}
\title{Wrapper function for fastx_barcode_splitter}
\usage{
fastx_barcode_splitter(
  fileTofxc,
  bcFile,
  mismatches = 0,
  fbs = "fastx_barcode_splitter.pl",
  stderr = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_barcode_splitter_stderr.txt")),
  stdout = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_barcode_splitter_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofxc}{path to file to process (fastq or fasta).}

\item{bcFile}{tab-delimited barcode file.}

\item{mismatches}{number of mismatches allowed.}

\item{fbs}{path to fastx_barcode_splitter.pl from FASTX toolkit.}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to split files; files will be written without any extension, in the same format as the input file.
}
\description{
Split multiplexed samples by user-defined indics using fastx_barcode_splitter
}
\author{
Kathryn Rozen-Gagnon
}
