% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_clipper}
\alias{fastx_clipper}
\title{Wrapper function for fastx_clipper}
\usage{
fastx_clipper(
  fileTofqs,
  outFile = paste0(file_path_sans_ext(fileTofqs), "_clip.", file_ext(fileTofqs)),
  fqc = "fastx_clipper",
  length = 18,
  adapter = "GTGTCAGTCACTTCCAGCGG",
  qEncoding = NULL,
  writelog = T,
  stderr = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_clipper_stderr.txt")),
  stdout = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_clipper_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqs}{path to file to process (fastq or fasta).}

\item{outFile}{output file path.}

\item{fqc}{path to fastx_clipper from FastX toolkit.}

\item{length}{miminum sequence length, default is 18 (set to .}

\item{adapter}{adapter to remove (default is "GTGTCAGTCACTTCCAGCGG", specify a string to change adapter sequence).}

\item{qEncoding}{quality encoding, set to either 33 (Sanger Phred+33 encoding/Illumina fastq) or NULL (Phred+64 fastq or fasta; default).}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE.}

\item{stderr}{path to stderr file.}

\item{stdout}{path to stdout file.}

\item{useClipRConda}{use conda environment installed by Herper, TRUE (default) or FALSE.}

\item{additional_Args}{additional arguments to be passed to system call.}

\item{verbose}{print messages to screen, TRUE or FALSE (default).}
}
\value{
path to clipped file
}
\description{
Wrapper function for fastx_clipper
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20, adapter = "GTGTCAG")
}
\author{
Kathryn Rozen-Gagnon
}
