% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_collapser}
\alias{fastx_collapser}
\title{Wrapper function for fastx_collapser}
\usage{
fastx_collapser(
  fileTofxc,
  outFile = file.path(dirname(fileTofxc), gsub("\\\\.fastq|\\\\.fq", "_collapse.fasta",
    basename(fileTofxc))),
  fxc = "fastx_collapser",
  qEncoding = 33,
  stderr = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_collapse_stderr.txt")),
  stdout = file.path(dirname(fileTofxc), paste0(basename(fileTofxc),
    "_fastx_collapse_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE,
  writelog = T
)
}
\arguments{
\item{fileTofxc}{File to process.}

\item{outFile}{Output file name.}

\item{fxc}{Path to fastx_collapser from FastX toolkit}

\item{qEncoding}{Quality encoding}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastx_collapser
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
}
\author{
Kathryn Rozen-Gagnon
}
