% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_quality_stats}
\alias{fastx_quality_stats}
\title{Wrapper function for fastx_quality_stats}
\usage{
fastx_quality_stats(
  fileTofqs,
  outFile = file.path(dirname(fileTofqs), gsub("\\\\.fastq|fq|fq\\\\.gz|fastq\\\\.gz",
    ".txt", basename(fileTofqs))),
  fqs = "fastx_quality_stats",
  qEncoding = 33,
  stderr = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_quality_stats_stderr.txt")),
  stdout = file.path(dirname(fileTofqs), paste0(basename(fileTofqs),
    "_fastx_quality_stats_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqs}{File to process.}

\item{outFile}{Output file name.}

\item{fqs}{Path to fastx_quality_stats from FastX toolkit}

\item{qEncoding}{Quality encoding}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print messages to screen}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for fastx_quality_stats
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
Fq_Stats <- fastx_quality_stats(FqFile_QF)
}
\author{
Kathryn Rozen-Gagnon
}
