% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_fastxtoolkit.R
\docType{methods}
\name{fastx_trimmer}
\alias{fastx_trimmer}
\title{Wrapper function for fastx_trimmer}
\usage{
fastx_trimmer(
  fileTofqt,
  fqt = "fastx_trimmer",
  read_start = 10,
  read_end = NULL,
  outFile = paste0(file_path_sans_ext(fileTofqt), "_trim.", file_ext(fileTofqt)),
  writelog = T,
  stderr = file.path(dirname(fileTofqt), paste0(basename(fileTofqt),
    "_trimmer_stats_stderr.txt")),
  stdout = file.path(dirname(fileTofqt), paste0(basename(fileTofqt),
    "_trimmer_stats_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqt}{File to process}

\item{fqt}{Path to fastx_trimmer from FastX toolkit}

\item{read_start}{read starting base}

\item{read_end}{read ending base}

\item{outFile}{Output file path}

\item{writelog}{write stderr/stdout logs, TRUE (default) or FALSE}

\item{stderr}{Path to stderr file}

\item{stdout}{Path to stdout file}

\item{useClipRConda}{Boolean on whether to use conda environment install by Herper}

\item{additionalArgumements}{Additional arguments to be passed to system call}

\item{verbose}{Print more message to screen}
}
\value{
Path to trimmed file
}
\description{
Wrapper function for fastx_trimmer
}
\examples{
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
}
\author{
Kathryn Rozen-Gagnon
}
