% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{fetchSequencesForClIP}
\alias{fetchSequencesForClIP}
\title{Retrieve sequences from under peaks/regions}
\usage{
fetchSequencesForClIP(
  peaks,
  resize = NULL,
  fasta,
  add5 = 0,
  add3 = 0,
  verbose = FALSE,
  bedHeader = TRUE,
  bedSep = "\\t"
)
}
\arguments{
\item{peaks}{ClIP peaks to process.}

\item{resize}{Size of window in bp for resizing around peak center.}

\item{fasta}{Genome fasta file.}

\item{add5}{Bp to add to 5' of resized peak.}

\item{add3}{Bp to add to 3' of resized peak.}

\item{verbose}{Print messages.}

\item{bedHeader}{TRUE if peak file contains column headers.}

\item{bedSep}{Seperator in BED file.}
}
\value{
Path to unzipped file.
}
\description{
Retrieve sequences from under peaks/regions
}
\examples{
 
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
decom <- decompress(testFQ,overwrite=TRUE)
com <- compress(decom,overwrite=TRUE)
}
\author{
Kathryn Rozen-Gagnon
}
