% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions_homer.R
\docType{methods}
\name{homer_peaks}
\alias{homer_peaks}
\title{Wrapper function for Homer's makeTagDirectory and findpeaks}
\usage{
homer_peaks(
  fileTofqs,
  maketagdir = "makeTagDirectory",
  findpeaks = "findpeaks",
  format = "bed",
  createSingleTagsTSV = TRUE,
  tagdir = file.path(dirname(fileTofqs), gsub("\\\\.bed", "",
    make.names(basename(fileTofqs)))),
  style = "factor",
  foldEnrichmentOverLocal = 2,
  localSize = 10000,
  strand = "seperate",
  minDist = 50,
  size = 10,
  fragLength = 10,
  stderr = paste0(getwd(), "homer_stats_stderr"),
  stdout = paste0(getwd(), "homer_stats_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fileTofqs}{File to process.}

\item{maketagdir}{Path to makeTagDirectory from Homer toolkit}

\item{findpeaks}{Path to findpeaks from Homer toolkit}

\item{format}{Input format}

\item{createSingleTagsTSV}{Create a single tags.tsv file for all "chromosomes"}

\item{tagdir}{Name of Tag Directory}

\item{style}{Type of Homer peak calling}

\item{foldEnrichmentOverLocal}{fold enrichment over local tag count}

\item{localSize}{region to check for local tag enrichment}

\item{strand}{find peaks using tags on both strands or separate}

\item{minDist}{minimum distance between peaks}

\item{size}{Peak size}

\item{fragLength}{Approximate fragment length,}

\item{stderr}{Path to stderr file.}

\item{stdout}{Path to stdout file.}

\item{useClipRConda}{Boolean on whether to use conda environment install by CondaSysReqs}

\item{additionalArgumements}{Additional arguments to be passed to system call.}

\item{verbose}{Print more message to screen.}
}
\value{
Path to unzipped file
}
\description{
Wrapper function for Homer's makeTagDirectory and findpeaks
}
\examples{
testFasta <- system.file("extdata/hg19Small.fa",package="clipR")
myIndex <- bowtie2_index(testFasta)
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile <- decompress(testFQ,overwrite=TRUE)
FqFile_QF <- fastq_quality_filter(FqFile)
FqFile_QFCollapsed <- fastx_collapser(FqFile_QF)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_QFCollapsed)
FqFile_QFColStpClipped <- fastx_clipper(FqFile_QFColStripped)
bam <- bowtie_align(FqFile_QFColStpClipped,myIndex)
bed <- bamtobed(bam)
homer_peaks(bed)
}
\author{
Kathryn Rozen-Gagnon
}
