% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{install_ctk}
\alias{install_ctk}
\title{Install ctk pipeline}
\usage{
install_ctk(path = NULL)
}
\arguments{
\item{path}{Path to where to install ctk and czplib}
}
\description{
Install ctk pipeline
}
\examples{
install_ctk()
getOption("CLIPflexR.condaEnv")
getOption("CLIPflexR.ctk")
getOption("CLIPflexR.czplib")
}
\author{
Kathryn Rozen-Gagnon Thomas Carroll Ji-Dung Luo
}
