% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{Ranges_count}
\alias{Ranges_count}
\title{Ranges_count}
\usage{
Ranges_count(
  fastas,
  miRNA_ranges,
  genomeIndex,
  linkers = NULL,
  length_max = NULL,
  length_min = NULL,
  bpparam = NULL,
  verbose = TRUE
)
}
\arguments{
\item{fastas}{FASTA files to process.}

\item{miRNA_ranges}{BED file containing annotated miRNA or other small RNA genomic ranges}

\item{genomeIndex}{full genome index}

\item{linkers}{contaminating sequences to remove (default is NULL)}

\item{length_max}{maximum length required during fasta processing (default is NULL)}

\item{length_min}{minimum length required during fasta processing (default is NULL)}

\item{bpparam}{TRUE or FALSE (default)}

\item{verbose}{print messages, TRUE (default) or FALSE}
}
\value{
Path
}
\description{
Map unprocessed or processed reads to genome and count small RNAs by location
}
\examples{
testFastq <- system.file("extdata/SRR1742056.fastq.gz",package="CLIPflexR")
}
\author{
Kathryn Rozen-Gagnon
}
