% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Revmap_count2.R
\docType{methods}
\name{revmap_count}
\alias{revmap_count}
\title{Reverse map small RNAs to processed reads}
\usage{
revmap_count(
  fastas,
  knownMiRNAs,
  bpparam = NULL,
  verbose = TRUE,
  linkers = NULL,
  length_max = NULL,
  length_min = NULL,
  removedups = FALSE
)
}
\arguments{
\item{fastas}{FASTA files to reverse map.}

\item{knownMiRNAs}{knownMiRNAs FASTA file.}

\item{linkers}{contaminating sequences to remove}
}
\value{
Path
}
\description{
Merge read sequence with bedfile
}
\author{
Kathryn Rozen-Gagnon
}
