% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{revmap_count}
\alias{revmap_count}
\title{Reverse map small RNAs to processed reads}
\usage{
revmap_count(
  fastas,
  knownMiRNAs,
  bpparam = NULL,
  verbose = TRUE,
  linkers = NULL,
  length_max = NULL,
  length_min = NULL,
  removedups = FALSE
)
}
\arguments{
\item{fastas}{FASTA files to reverse map}

\item{knownMiRNAs}{FASTA file containing annotated miRNA or other small RNA sequences}

\item{bpparam}{TRUE or FALSE (default)}

\item{verbose}{print messages, TRUE (default) or FALSE}

\item{linkers}{contaminating sequences to remove}

\item{length_max}{maximum length required during fasta processing, integer (default is NULL)}

\item{length_min}{minimum length required during fasta processing, integer (default is NULL)}

\item{removedups}{remove multiple miRNAs or small RNAs mapping to the same read, TRUE or FALSE (FALSE)}
}
\value{
BEDs containing counts of
}
\description{
Reverse map small RNAs to processed reads
}
\author{
Kathryn Rozen-Gagnon
}
