% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{revmap_process}
\alias{revmap_process}
\title{Revmap_process}
\usage{
revmap_process(fastas, linkers = NULL, length_max = NULL, length_min = NULL)
}
\arguments{
\item{fastas}{FASTA file to process}

\item{linkers}{contaminating seqences to remove (default is NULL)}

\item{length_max}{maximum length required during fasta processing, integer (default is NULL)}

\item{length_min}{minimum length required during fasta processing, integer (default is NULL)}
}
\value{
Processed FASTAS
}
\description{
Process reads for small RNA counting
}
\author{
Kathryn Rozen-Gagnon
}
