% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Revmap_count2.R, R/miRNAs_by_ranges.R
\docType{methods}
\name{revmap_process}
\alias{revmap_process}
\title{Process reads for miRNA counting}
\usage{
revmap_process(input, linkers = NULL, length_max = NULL, length_min = NULL)

revmap_process(input, linkers = NULL, length_max = NULL, length_min = NULL)
}
\arguments{
\item{linkers}{contaminating seqences to remove}

\item{fasta}{FASTA file to process}

\item{readlength}{length requirements for FASTA}

\item{outfa}{PATH to output FASTA}
}
\value{
Path to FASTA file

Path to FASTA file
}
\description{
Process reads for miRNA counting

Process reads for miRNA counting
}
\examples{
bam <- system.file("extdata/example_bams/KRG121817A_Aag2_Ago1_Exp1A_rm5_rm3.bam",package="CLIPflexR")
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
unbam(bam)
bam <- system.file("extdata/example_bams/KRG121817A_Aag2_Ago1_Exp1A_rm5_rm3.bam",package="CLIPflexR")
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="clipR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)

}
\author{
Kathryn Rozen-Gagnon

Kathryn Rozen-Gagnon
}
