% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{methods}
\name{revmap_process}
\alias{revmap_process}
\title{Process reads for small RNA counting}
\usage{
revmap_process(fastas, linkers = NULL, length_max = NULL, length_min = NULL)
}
\arguments{
\item{fastas}{FASTA file to process}

\item{linkers}{contaminating seqences to remove (default is NULL)}

\item{length_max}{maximum length required during fasta processing, integer (default is NULL)}

\item{length_min}{minimum length required during fasta processing, integer (default is NULL)}
}
\value{
Processed FASTAS
}
\description{
Process reads for small RNA counting
}
\examples{
bam <- system.file("extdata/xxbam",package="CLIPflexR")
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)
FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)
FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
unbam(bam)
}
\author{
Kathryn Rozen-Gagnon
}
